Reorganize type declaration
