+++ cmd/start.go
@@ -21,6 +21,7 @@ func startCmd() *cobra.Command {
 				// nolint
 				log.Fatalf("Failed to load config: %v", err)
 			}
+
 			runDAG(cmd.Context(),
 				cfg,
 				engine.New(
