+++ internal/dag/dag.go
@@ -14,46 +14,6 @@ import (
 	"github.com/robfig/cron/v3"
 )
 
-// Schedule contains the cron expression and the parsed cron schedule.
-type Schedule struct {
-	// Expression is the cron expression.
-	Expression string `json:"Expression"`
-	// Parsed is the parsed cron schedule.
-	Parsed cron.Schedule `json:"-"`
-}
-
-// HandlerOn contains the steps to be executed on different events in the DAG.
-type HandlerOn struct {
-	Failure *Step `json:"Failure"`
-	Success *Step `json:"Success"`
-	Cancel  *Step `json:"Cancel"`
-	Exit    *Step `json:"Exit"`
-}
-
-// MailOn contains the conditions to send mail.
-type MailOn struct {
-	Failure bool `json:"Failure"`
-	Success bool `json:"Success"`
-}
-
-// SMTPConfig contains the SMTP configuration.
-type SMTPConfig struct {
-	Host     string `json:"Host"`
-	Port     string `json:"Port"`
-	Username string `json:"Username"`
-	Password string `json:"Password"`
-}
-
-// MailConfig contains the mail configuration.
-type MailConfig struct {
-	From string `json:"From"`
-	To   string `json:"To"`
-	// Prefix is the prefix for the subject of the mail.
-	Prefix string `json:"Prefix"`
-	// AttachLogs is the flag to attach the logs in the mail.
-	AttachLogs bool `json:"AttachLogs"`
-}
-
 // DAG contains all information about a workflow.
 type DAG struct {
 	// Location is the absolute path to the DAG file.
@@ -129,6 +89,46 @@ type DAG struct {
 	HistRetentionDays int `json:"HistRetentionDays"`
 }
 
+// Schedule contains the cron expression and the parsed cron schedule.
+type Schedule struct {
+	// Expression is the cron expression.
+	Expression string `json:"Expression"`
+	// Parsed is the parsed cron schedule.
+	Parsed cron.Schedule `json:"-"`
+}
+
+// HandlerOn contains the steps to be executed on different events in the DAG.
+type HandlerOn struct {
+	Failure *Step `json:"Failure"`
+	Success *Step `json:"Success"`
+	Cancel  *Step `json:"Cancel"`
+	Exit    *Step `json:"Exit"`
+}
+
+// MailOn contains the conditions to send mail.
+type MailOn struct {
+	Failure bool `json:"Failure"`
+	Success bool `json:"Success"`
+}
+
+// SMTPConfig contains the SMTP configuration.
+type SMTPConfig struct {
+	Host     string `json:"Host"`
+	Port     string `json:"Port"`
+	Username string `json:"Username"`
+	Password string `json:"Password"`
+}
+
+// MailConfig contains the mail configuration.
+type MailConfig struct {
+	From string `json:"From"`
+	To   string `json:"To"`
+	// Prefix is the prefix for the subject of the mail.
+	Prefix string `json:"Prefix"`
+	// AttachLogs is the flag to attach the logs in the mail.
+	AttachLogs bool `json:"AttachLogs"`
+}
+
 // HandlerType is the type of the handler.
 type HandlerType string
 
@@ -155,6 +155,11 @@ var (
 		"onCancel":  HandlerOnCancel,
 		"onExit":    HandlerOnExit,
 	}
+)
+
+var (
+	defaultHistoryRetentionDays = 30
+	defaultMaxCleanUpTime       = time.Second * 60
 )
 
 // setup sets the default values for the DAG.
